Blog-list 4.21

+++ part4/blog_list/controllers/blogs.js
@@ -1,3 +1,4 @@
+/* eslint-disable consistent-return */
 const jwt = require('jsonwebtoken');
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
@@ -43,6 +44,22 @@ blogsRouter.post('/', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const { token } = request;
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({
+      error: 'token missing or invalid',
+    });
+  }
+
+  const user = await User.findById(decodedToken.id);
+  const blog = await Blog.findById(request.params.id);
+  if (blog.user.toString() !== user.id.toString()) {
+    return response.status(401).json({
+      error: 'cannot delete other users blogs',
+    });
+  }
+
   await Blog.findByIdAndRemove(request.params.id);
   response.status(204).end();
 });
